requirements.txt
Click==7.0
gitdb2==2.0.6
GitPython==3.0.5
munch==2.5.0
six==1.14.0
smmap2==2.0.5
